[MINDEXER-202] Modularize
Rework, package changes in Search API to not share
packages, and modularize.

---

https://issues.apache.org/jira/browse/MINDEXER-202
