+++ indexer-core/src/main/java/module-info.java
@@ -1,0 +1,16 @@
+module indexer.core {
+    exports org.apache.maven.index;
+    exports org.apache.maven.index.archetype;
+    exports org.apache.maven.index.artifact;
+    exports org.apache.maven.index.context;
+    exports org.apache.maven.index.creator;
+    exports org.apache.maven.index.expr;
+    exports org.apache.maven.index.fs;
+    exports org.apache.maven.index.incremental;
+    exports org.apache.maven.index.locator;
+    exports org.apache.maven.index.packer;
+    exports org.apache.maven.index.search.grouping; // unfortunate
+    exports org.apache.maven.index.treeview;
+    exports org.apache.maven.index.updater;
+    exports org.apache.maven.index.util;
+}
