+++ search-api/src/main/java/module-info.java
@@ -1,0 +1,6 @@
+module search.api {
+    exports org.apache.maven.search.api;
+    exports org.apache.maven.search.api.request;
+    exports org.apache.maven.search.api.support;
+    exports org.apache.maven.search.api.transport;
+}
