+++ search-api/src/main/java/org/apache/maven/search/api/support/SearchResponseSupport.java
@@ -16,13 +16,13 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.maven.search.support;
+package org.apache.maven.search.api.support;
 
 import java.util.List;
 
-import org.apache.maven.search.Record;
-import org.apache.maven.search.SearchRequest;
-import org.apache.maven.search.SearchResponse;
+import org.apache.maven.search.api.Record;
+import org.apache.maven.search.api.SearchRequest;
+import org.apache.maven.search.api.SearchResponse;
 
 import static java.util.Objects.requireNonNull;
