+++ search-api/src/main/java/org/apache/maven/search/api/transport/Transport.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.maven.search.transport;
+package org.apache.maven.search.api.transport;
 
 import java.io.Closeable;
 import java.io.IOException;
