+++ search-backend-indexer/src/main/java/module-info.java
@@ -1,0 +1,5 @@
+module search.backend.indexer {
+    requires search.api;
+    requires indexer.core;
+    exports org.apache.maven.search.backend.indexer;
+}
