+++ search-backend-remoterepository/src/main/java/module-info.java
@@ -1,0 +1,4 @@
+module search.backend.remoterepository {
+    requires search.api;
+    exports org.apache.maven.search.backend.remoterepository;
+}
