+++ search-backend-remoterepository/src/main/java/org/apache/maven/search/backend/remoterepository/RemoteRepositorySearchResponse.java
@@ -18,7 +18,7 @@
  */
 package org.apache.maven.search.backend.remoterepository;
 
-import org.apache.maven.search.SearchResponse;
+import org.apache.maven.search.api.SearchResponse;
 import org.jsoup.nodes.Document;
 
 /**
