+++ search-backend-remoterepository/src/test/java/org/apache/maven/search/backend/remoterepository/internal/RemoteRepositorySearchBackendImplTest.java
@@ -26,15 +26,15 @@ import java.util.Collection;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
 
-import org.apache.maven.search.MAVEN;
-import org.apache.maven.search.Record;
-import org.apache.maven.search.SearchRequest;
+import org.apache.maven.search.api.MAVEN;
+import org.apache.maven.search.api.Record;
+import org.apache.maven.search.api.SearchRequest;
 import org.apache.maven.search.backend.remoterepository.RemoteRepositorySearchBackend;
 import org.apache.maven.search.backend.remoterepository.RemoteRepositorySearchBackendFactory;
 import org.apache.maven.search.backend.remoterepository.RemoteRepositorySearchResponse;
-import org.apache.maven.search.request.BooleanQuery;
-import org.apache.maven.search.request.FieldQuery;
-import org.apache.maven.search.request.Query;
+import org.apache.maven.search.api.request.BooleanQuery;
+import org.apache.maven.search.api.request.FieldQuery;
+import org.apache.maven.search.api.request.Query;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
