+++ search-backend-smo/src/main/java/module-info.java
@@ -1,0 +1,4 @@
+module search.backend.smo {
+    requires search.api;
+    exports org.apache.maven.search.backend.smo;
+}
