+++ search-backend-smo/src/main/java/org/apache/maven/search/backend/smo/SmoSearchBackend.java
@@ -20,8 +20,8 @@ package org.apache.maven.search.backend.smo;
 
 import java.io.IOException;
 
-import org.apache.maven.search.SearchBackend;
-import org.apache.maven.search.SearchRequest;
+import org.apache.maven.search.api.SearchBackend;
+import org.apache.maven.search.api.SearchRequest;
 
 /**
  * The SMO search backend.
